Moved the selection mechanic to a helper function, included the autosuggest for utilising the new selection method
