+++ src/aims_ui/static/js/e_all_pages/save_and_restore_input_helpers/autosuggest/autosuggest_restorers.mjs
@@ -1,14 +1,8 @@
-
+import { selectInputFromHtmlIdUsingContianer } from "../component_selection.mjs";
 
 export function restoreValuesToAutosuggestInput(page_name, htmlId, pagePreviouslySearchedValues) {
   // Get the input element
-
-  const inputElement = document.querySelector('#' + htmlId);
-
-  if (!inputElement) {
-    console.warn('No input element found for Id:', htmlId);
-    return;
-  }
+  const inputElement = selectInputFromHtmlIdUsingContianer(htmlId);
 
   // Get previous value
   const previousValue = pagePreviouslySearchedValues[htmlId];
