+++ src/aims_ui/static/js/e_all_pages/save_and_restore_input_helpers/component_selection.mjs
@@ -1,3 +1,4 @@
+// Functions to select components through use of the container divs
 
 export function selectInputFromHtmlIdUsingContianer(htmlId) {
   // Get the input container
@@ -5,7 +6,7 @@ export function selectInputFromHtmlIdUsingContianer(htmlId) {
   const containerElement = document.querySelector(containerQuerySelector);
 
   if (!containerElement) {
-    console.warn('No container element found for input Id:', containerQuerySelector);
+    console.warn('No container element found for input Id:' + containerQuerySelector + '. Have you ensured all components have a wrapper div with the ID "complete-container-for-[htmlId]"?');
     return null;
   }
